Move repository implementations to `types` module
