+++ packages/ploys/src/repository/types/fs/mod.rs
@@ -8,7 +8,9 @@ use walkdir::WalkDir;
 
 pub use self::error::Error;
 
-use super::{Commit, Repository, Stage, Staged};
+use crate::repository::{Commit, Repository, Stage};
+
+use super::staged::Staged;
 
 /// A file system repository.
 #[derive(Clone)]
