+++ packages/ploys/src/repository/types/mod.rs
@@ -1,0 +1,11 @@
+pub mod staged;
+pub mod staging;
+
+#[cfg(feature = "fs")]
+pub mod fs;
+
+#[cfg(feature = "git")]
+pub mod git;
+
+#[cfg(feature = "github")]
+pub mod github;
